correct use of uint32 for surviving corruptions

+++ message.go
@@ -147,7 +147,7 @@ func decodeMessage(packet []byte, payloadCodecsMap map[byte]PayloadCodec) (uint3
     msg.compression = byte(0)
     copy(msg.checksum[:], packet[5:9])
     payloadLength := length - 1 - 4
-    if len(packet) < 9+payloadLength {
+    if uint32(len(packet)) < 9+payloadLength {
       log.Printf("length mismatch in msg.magic == 0, expected at least: %X, was: %X\n", 9+payloadLength, len(packet))
       return 0, nil
     }
@@ -156,7 +156,7 @@ func decodeMessage(packet []byte, payloadCodecsMap map[byte]PayloadCodec) (uint3
     msg.compression = packet[5]
     copy(msg.checksum[:], packet[6:10])
     payloadLength := length - NO_LEN_HEADER_SIZE
-    if len(packet) < 10+payloadLength {
+    if uint32(len(packet)) < 10+payloadLength {
       log.Printf("length mismatch in msg.magic == MAGIC_DEFAULT, expected at least: %X, was: %X\n", 10+payloadLength, len(packet))
       return 0, nil
     }
